Initial import of codes
This commit mostly copies over code from where we originally wrote it. I
made some small adjustments, like restructuring the directories and
wiring up the import paths correctly, but for the most part this is a
straight up copypasta.

There's still plenty left to do, like set up testing, eslint, and babel,
but I'll leave that for a later commit.