Move documentation from withStyles.jsx to readme
We need a readme so people know what this thing is and how to use it.
I've already written some documentation in the withStyles.jsx file that
we can use as a starting-point. This commit mostly just copies it over
as-is with some minor formatting adjustments. Future commits will shape
it into its final form.