Add React Native interface
React Native does most of the heavy lifting for us, so this interface is
very simple.

Requiring an unbuilt react-native results in some issues. First I ran
into __DEV__ not being defined. I was able to cross this hurdle by
defining these global variables, but then I ran into a different issue
so I decided to take a different approach.

To make the tests work I brought in react-native-mock. This package
mocks out StyleSheet.create in a way that makes it convenient to write
assertions on the styles, but it obscures the true result of the
underlying implementation in a way that I don't want here. So, I've come
up with a workaround that lets us have tests for this without too much
trouble.

https://github.com/lelandrichardson/react-native-mock/issues/1

Ideally, we would be testing this with react-native directly, but I
think this is an alright compromise for now.