Extract interfaces to their own repos
It turns out that optionalDependencies doesn't work the way I expected
it to. It will install all of the optionalDependencies unless
`--no-optional` is used all the time by consumers when running `npm
install`. This is no good because people not using react-native or
aphrodite shouldn't have to install those dependencies.

So, I extracted the interfaces to their own repos, published with their
own packages.

- https://github.com/airbnb/react-with-styles-interface-aphrodite
- https://github.com/airbnb/react-with-styles-interface-react-native