Avoid copying withStyles propTypes from wrapped component
In 094595b64 we started copying propTypes and defaultProps from the
wrapped component up to the wrapping component. Unfortunately, we forgot
about the two props that are provided by the wrapping component (styles
and theme), which may end up in the propTypes of the wrapped component.
This caused propTypes warnings to happen.

Since these props are provided by this component, we don't expect them
to be passed in by consumers, so let's just delete them. It would be
nice if there was an easy way to forbid them at the propType level, but
I'll leave that for another time.