Add pureComponent option
By default `withStyles()` will create a component that extends
`React.Component`. If you want to apply the `shouldComponentUpdate()`
optimization offered by `React.PureComponent`, you can set the
`pureComponent` option to `true`.

Obviously, you could choose to compose this HOC with `pure` from
recompose or something like it, but since this is such a small change to
make, it seems valuable enough to offer it for convenience.

This will only work if you are using React 15.3.0 or later.