Fix tests
I noticed that in pretty much all of the tests where we had our
expectation inside render of the wrapped component, we weren't actually
running the exception because we were shallow rendering. To fix this, I
am adding some .dive() calls in a few places which was designed for this
use-case. This uncovered a couple of bugs in the tests, which I also
addressed in this commit.