Defer StyleSheet creation to withStyles' `componentWillMount` method
After some offline conversation about the best approach to handling
specificity issues in react-with-styles-interface-aphrodite/with-rtl, we
arrived at a three step approach that would eliminate specificity issue
entirely.

1. In order to be able to modify built-in styles based on context,
defer stylesheet creation to the constructor of the withStyles HOC.
This will also give us the performance benefit of not calling
StyleSheet.create for currently unrendered components (but will
also slow down the first render a little bit)

2. Based on directional context, cache and create the appropriate
 StyleSheet in the constructor of the withStyles HOC.

3. Pass a directional css method down to the wrapped component as a
prop. This will be a breaking change and will require that components
use props.css instead of the react-with-styles export. We will initially
 address this change internally using a babel plugin and eventually do a
 sweeping codemod. The directional css method will flip or not flip
inline styles depending on context --- this will necessitate an update
to RWS-IA.

This is step 1 of the process.