Use performance.mark() and performance.measure()
It would be nice to have more insight into how react-with-styles is
performing in real world applications. We can use performance.mark() and
performance.measure() to annotate the user timing pane in Chrome's
performance devtool. I am using the same method React uses to allow this
code to be built out in production builds.

I wanted to give this its own style next to the React annotations, so I
decided to bring in an emoji: 👩‍🎨

I wanted to pass the component name down to the resolve() annotations,
but in my profiling I never actually saw that annotation appear since it
is so fast, and passing it down there would be a pain and potentially a
small performance hit, so I decided against that.

In this process, I think I found a bug in Chrome:

  https://bugs.chromium.org/p/chromium/issues/detail?id=837863