Remove componentWillMount
This method is deprecated in React 16.3. To prepare us for the new
world, I am removing our usage of it.

I initially thought that I should move the maybeCreateStyles() function
call into the constructor, but upon closer investigation, I'm not really
sure why this is called in componentWillMount at all, only to be called
again in render. My hunch is that this was done to make sure that the
styles were created before calling flush, but we can accomplish that by
reordering the method calls in render itself.