Move work out of render and into state
I noticed that we have been storing the direction in state, and then in
componentWillMount and render we have a bunch of logic that derifes the
styles and the resolve method from that bit of direction state. However,
these things should only change when the direction changes, so we should
be able to reduce the amount of work that needs to be done in render by
instead storing these values in state. This should improve performance
whenever a component wrapped with this HOC re-renders.

This also consolidates some of the logic, which I think makes the code a
little easier to follow.