Replace deepmerge with object.assign
I don't see any compelling reason to keep using deepmerge here. At most
we simply need to do a shallow copy of these objects, which
Object.assign is perfect for. To keep maximum compatibility, I decided
to bring in the object.assign package for this, which defaults to the
native Object.assign.