[withStyles] Conditionally forward ref to inner component
React.forwardRef was added in React 16.3, but since this package still
supports older versions of React, we conditionally fork the application
of the forwarded ref. We will remove the fork in favor of always
forwarding the ref in a later breaking change version.

context: https://github.com/airbnb/react-with-styles/pull/240#discussion_r533497857